REFACTOR : extract WebViewLayout() to their own file
Previously, WebViewLayout() is inside the MainActivity.kt

This commit extract WebViewLayout function to their own class in WebViewLayout.kt.

Thus, the file is more readable and every file have clear goals
